backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
python-multipart==0.0.9
httpx==0.28.1
jinja2==3.1.5
markdown-it-py==3.0.0
Pillow==11.1.0
playwright==1.50.0

pytest==8.3.4
